Implementing Selection Function class
This version has a object class that can accomodate two different selection
functions: the separable and the non-separable radial and angular parts. This
class would allow to treat the selection effect in the main code in a consistent
way, independently of how the selection function is determined. So far we have
implemented a constructor and a destructor and the data structure inside the class.
The memory allocation was tested and is working ok. Next step is to implement the
() operator so we can acess the selection function for field i and pixel j.

We also made the LoadTable function more robust: it accepts empty lines and comments.
We also implemented a LoadVecs function that is also robust.

+++ src/concatalm.sh
@@ -1,3 +1,4 @@
+alm2cl=$1 # 1 - Calculates Cl and erases alms; 0 - Keep alms and do not calculate Cls.
 
 prefix=recov-alm 
 # Join multiple realizations into one file:
@@ -12,11 +13,11 @@ echo "Erasing temp files from $prefix..."
 rm $prefix-*.dat -f
 rm ${prefix}_*.dat -f
 # Run python script to compute C(l)s:
-echo "Cling for $prefix..."
-../bin/alm2Cl.py $prefix.dat > ${prefix}_cl.dat
-rm $prefix.dat -f
-
-exit 0
+if [ $alm2cl -eq 1 ]; then
+    echo "Cling for $prefix..."
+    ../bin/alm2Cl.py $prefix.dat > ${prefix}_cl.dat
+    rm $prefix.dat -f
+fi
 
 prefix=aux-alm
 # Join multiple realizations into one file:
@@ -31,6 +32,8 @@ echo "Erasing temp files from $prefix..."
 rm $prefix-*.dat -f
 rm ${prefix}_*.dat -f
 # Run python script to compute C(l)s:
-echo "Cling for $prefix..."
-../bin/alm2Cl.py $prefix.dat > ${prefix}_cl.dat
-rm $prefix.dat -f
+if [ $alm2cl -eq 1 ]; then
+    echo "Cling for $prefix..."
+    ../bin/alm2Cl.py $prefix.dat > ${prefix}_cl.dat
+    rm $prefix.dat -f
+fi
